Types

Abstact Types

The abstract types below can be extended to add new functionality.

RetirementPlanners.AbstractGBM — Type

AbstractGBM <: ContinuousUnivariateDistribution

Abstract type for simulating investment growth using Geometric Brownian Motion.

RetirementPlanners.AbstractModel — Type

AbstractModel

An abstract model type for simulating retirement investments.

RetirementPlanners.AbstractState — Type

AbstractState

An abstract type for tracking the state of the model during simulation.

RetirementPlanners.AbstractLogger — Type

AbstractLogger

An abstract type for recording the state of the model during simulation.

Concrete Types

RetirementPlanners.GBM — Type

GBM{T<:Real} <: AbstractGBM

A distribution object for Geometric Brownian Motion (GBM), which is used to model growth of stocks.

Fields

μ::T: growth rate

σ::T: volitility in growth rate

x0::T: initial value of stock

x::T: current value

RetirementPlanners.VarGBM — Type

VarGBM{T<:Real} <: AbstractGBM

A distribution object for variable Geometric Brownian Motion (vGBM), which is used to model growth of stocks. Unlike GBM, vGBM selects growth rate (μ) and volitility (σ) parameters from a normal distribution on each simulation run to capture uncertainy in these parameters.

Fields

μ::T: growth rate sampled from normal distribution

σ::T: volitility in growth rate sampled from truncated normal distribution

αμ::T: mean of growth rate distribution

ασ::T: mean of volitility of growth rate distribution

ημ::T: standard deviation of growth rate distribution

ησ::T: standard deviation of volitility of growth rate distribution

x0::T: initial value of stock

x::T: current value

Constructor

VarGBM(; αμ, ασ, ημ, ησ, x0=1.0, x=x0)

RetirementPlanners.MvGBM — Type

MvGBM{T<:Real} <: AbstractGBM

A distribution object for Multivariate Geometric Brownian Motion (MvGBM), which is used to model growth of multiple stocks and bonds.

Fields

μ::Vector{T}: growth rates

σ::Vector{T}: volitility in growth rates

x0::Vector{T}: initial value of stocks

x::Vector{T}: value of stocks

ratios::Vector{T}: allocation proportion of stocks/bonds

ρ::Array{T,2}: correlation matrix between stocks/bonds

Σ::Array{T,2}: covariance matrix between stocks/bonds

Constructor

MvGBM(; μ, σ, ρ, ratios)

RetirementPlanners.Model — Type

Model{S} <: AbstractModel

The default retirement simulation Model.

Fields

Δt::Float64: the time step of the simulation in years

duration::Float64: the duration of the simulation in years

start_age::Float64: age at the beginning of the simulation

start_amount::Float64: initial investment amount

state::S: the current state of the system

withdraw!: a function called on each time step to withdraw from investments

invest!: a function called on each time step to invest money into investments

update_income!: a function called on each time step to update income sources

update_inflation!: a function called on each time step to compute inflation

update_interest!: a function called on each time step to compute interest on investments

update_net_worth!: a function called on each time step to compute net worth

log!: a function called on each time step to log data

Constructor

Model(;
        Δt,
        duration,
        start_age,
        start_amount,
        state = State(),
        withdraw! = variable_withdraw,
        invest! = variable_invest,
        update_income! = fixed_income,
        update_inflation! = dynamic_inflation,
        update_interest! = dynamic_interest,
        update_net_worth! = default_net_worth,
        log! = default_log!
    )

RetirementPlanners.State — Type

State <: AbstractState

Represents the state of the model, which is updated on each iteration.

Fields

interest_rate::Float64: interest rate of investment during the current time period

inflation_rate::Float64: the inflation rate during the current time period

income_amount::Float64: income during the current time period from various sources, e.g., social security, pension, etc.

invest_amount::Float64: the amount invested during the current time period

withdraw_amount::Float64: the amount deducted from investments during the current time period

net_worth::Float64: total value of the investment during the current time period

RetirementPlanners.Logger — Type

Logger{T<:Real} <: AbstractLogger

An object for storing variables of the simulation.

Fields

net_worth::Array{T,2}: total value of investments

interest::Array{T,2}: growth rate of investment

inflation::Array{T,2}: inflation rate

In each array above, rows are time steps and columns are repetitions of the simulation.

Methods

General Methods

RetirementPlanners.get_times — Function

get_times(model::AbstractModel)

Returns the time steps used in the simulation.

Arguments

model::AbstractModel: an abstract Model object

RetirementPlanners.grid_search — Function

grid_search(
    model::AbstractModel,
    Logger::Type{<:AbstractLogger},
    n_reps; 
    config::NamedTuple,
    threaded::Bool = false,
    n = 1,
    show_progress::Bool = false,
    yoked_values = (),
    kwargs...,
)

Performs a grid search over vectorized inputs specified in the configuration setup. As an example, consider the following configuration setup:

config = (
    # withdraw parameters 
    kw_withdraw = (
        distribution = [
            Normal(3000, 1000), 
            Normal(4000, 1000),
        ],
        start_age = 65,
    ),
    # invest parameters
    kw_invest = (
        distribution = [
            Normal(1000, 100),
            Normal(1500, 100),
        ]
        end_age = 65,
    ),
    # interest parameters
    kw_interest = (
        gbm = GBM(; μ = .07, σ = .05),
    ),
    # inflation parameters
    kw_inflation = (
        gbm = GBM(; μ = .035, σ = .005),
    )
)

In the example above, four simulations will be performed: one for each combination of withdraw distribution and investment distribution.

Arguments

model::AbstractModel: an abstract model type for performing Monte Carlo simulations of investment scenarios

Logger::Type{<:AbstractLogger}: a type for collecting variables of the simulation. The constructor signature is Logger(; n_reps, n_steps)

n_reps: the number of times the investiment simulation is repeated for each input combination.

Keywords

config::NamedTuple: a nested configuration of the simulation parameters.

threaded::Bool = false: runs simulations on separate threads if true

show_progress::Bool = false: shows progress bar if true

yoked_values = (): fix specified inputs to have the same values, such as: [Pair((:kw_withdraw, :start_age), (:kw_invest, :end_age))]

Output

Returns a vector of tuples where each tuple corresponds to the result of a single simulation condition. Each tuple consists of input values and output results. The function to_dataframe can be used to transform output into a long-form DataFrame.

Notes

This function was inspired by parmscan in Agents.jl.

RetirementPlanners.is_event_time — Function

is_event_time(model::AbstractModel, t, rate)

Indicate whether it is time for a periodic event to occur.

Arguments

model::AbstractModel: an abstract Model object

t: current time (or age) in years

rate: the interval between repeating events measured in years

Base.rand — Function

rand(dist::AbstractGBM, n_steps, n_reps; Δt)

Simulate a random trajector of a Geometric Brownian motion process.

Arguments

dist::AbstractGBM: a distribution object for Geometric Brownian Motion

n_steps: the number of discrete time steps in the simulation

n_reps: the number of times the simulation is repeated

Keywords

Δt: the time step for Geometric Brownian Motion

RetirementPlanners.simulate! — Function

simulate!(
    model::AbstractModel,
    logger::AbstractLogger,
    n_reps;
    kwargs...
)

Simulate the a retirement scenario a specified number of times.

Arguments

model::AbstractModel: an abstract Model object

logger::AbstractLogger: an object for storing variables of the simulation

n_reps: the number of times to repeat the Monte Carlo simulation

Keywords

kwargs...: optional keyword arguments passed to update!

RetirementPlanners.update! — Function

update!(
    model::AbstractModel,
    logger::AbstractLogger,
    step,
    rep,
    t; 
    kw_income=(),
    kw_withdraw=(),
    kw_invest=(),
    kw_inflation=(),
    kw_interest=(),
    kw_net_worth=(),
    kw_log=()
)

Performs an update on each time step by calling the following functions defined in model:

update_income!: update sources of income, such as social security, pension etc.

withdraw!: withdraw money

invest!: invest money

update_inflation!: compute inflation

update_interest!: compute interest

update_net_worth!: compute net worth for the time step

log!: log desired variables

Each function except log! has the signature my_func(model, t; kwargs...). The function log! has the signature log!(model, logger, step, rep; kwargs...).

Arguments

model::AbstractModel: an abstract Model object

rep::Int: repetition count of simulation

time_step::Int: time step count of simulation

t: time in years

Keywords

kw_income = (): optional keyword arguments passed to update_income!

kw_withdraw = (): optional keyword arguments passed to withdraw!

kw_invest = (): optional keyword arguments passed to invest!

kw_inflation = (): optional keyword arguments passed to update_inflation!

kw_interest = (): optional keyword arguments passed to update_interest!

kw_net_worth = (): optional keyword arguments passed to update_net_worth!

kw_log = (): optional keyword arguments passed to log!

Update Methods

Update Income

RetirementPlanners.fixed_income — Function

fixed_income(
    model::AbstractModel,
    t;
    social_security_income = 0.0,
    pension_income = 0.0,
    social_security_start_age = 67.0,
    pension_start_age = 67
)

Recieve a fixed amount of income (e.g., social security, pension) per time step

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

social_security_income = 0.0: income from social security on a per period basis

pension_income = 0.0: income from pensions on a per period basis

social_security_start_age = 67.0: age at which social security income begins

pension_start_age = 67: age at which pension income begins

Update Inflation

RetirementPlanners.fixed_inflation — Function

fixed_inflation(
    model::AbstractModel, 
    t;
    inflation_rate = .03
)

Returns a fixed inflation rate of a specified value.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

inflation_rate = .03: a constant rate of inflation per year

RetirementPlanners.variable_inflation — Function

variable_inflation(
    model::AbstractModel,
    t;
    distribution = Normal(.03, .01)
)

Returns an interest rate sampled from a specified distribution.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

distribution = Normal(.03, .01): the distribution of inflation per year

RetirementPlanners.dynamic_inflation — Function

dynamic_inflation(
    model::AbstractModel, 
    t; 
    gbm = GBM(; μ=.03, σ=.01, x0=1)
)

Models inflation in the stock market as a geometric brownian motion process.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keyword

gbm = GBM(; μ=.03, σ=.01, x0=1): a geometric brownian motion object with parameters

μ reflecting mean growth rate, and σ reflecting volitility in growth rate. The parameter x0 sets an arbitrary scale. The function also supports VarGBM.

Update Interest

RetirementPlanners.fixed_interest — Function

fixed_interest(model::AbstractModel, t; interest_rate = .07)

Returns a fixed interesting rate using a specified value.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

interest_rate = .07: a constant rate of investment growth per year

RetirementPlanners.variable_interest — Function

variable_interest(
    model::AbstractModel,
    t;
    distribution = Normal(.07, .05)
)

Returns interest rate sampled from a specified distribution.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

distribution = Normal(.07, .05): the distribution of interest per year

RetirementPlanners.dynamic_interest — Function

dynamic_interest(
    model::AbstractModel,
    t; 
    gbm = GBM(; μ=.07, σ=.05, x0=1)
)

Models interest in the stock market as a geometric brownian motion process.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keyword

gbm = GBM(; μ=.07, σ=.05, x0=1): a geometric Brownian motion object with parameters

μ reflecting mean growth rate, and σ reflecting volitility in growth rate. The parameter x0 sets an arbitrary scale. Other variations of geometric Brownian motion can be used, including VarGBM and MvGBM

rebalance_rate = Inf: the time elapsed in years between rebalacing the portfolio. Not applicable

to GBM

Update Investments

RetirementPlanners.fixed_investment — Function

fixed_investment(
    model::AbstractModel, 
    t;
    invest_amount = 1000.0,
    end_age = 67.0
)

Contribute a fixed amount into investments per time step.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

invest_amount = 3000.0: the amount contributed to investments per time step

start_age = 67.0: the age at which investing stops

RetirementPlanners.variable_investment — Function

variable_investment(
    model::AbstractModel,
    t; 
    end_age = 67.0, 
    distribution = Normal(1000, 200)
)

Contribute a variable amount into investments per time step using the specifed distribution.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

end_age = 67.0: the age at which investing stops

distribution = Normal(1000, 200): the distribution from which the investment amount is sampled

Log

RetirementPlanners.default_log! — Function

default_log!(
    model::AbstractModel,
    logger,
    step,
    rep;
    _...
)

Logs the following information on each time step of each simulation repetition:

net worth

interest rate

inflation rate

Arguments

model::AbstractModel: an abstract Model object

logger: a logger object

Update Net Worth

RetirementPlanners.default_net_worth — Function

default_net_worth(model::AbstractModel, t)

Computes net worth for the current time step as follows:

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Update Withdraw

RetirementPlanners.adaptive_withdraw — Function

adaptive_withdraw(
    model::AbstractModel,
    t;
    start_age = 67, 
    min_withdraw = 1000,
    percent_of_real_growth = 1,
    income_adjustment = 0.0,
    volitility = .1,
)

An adaptive withdraw scheme based on current real growth rate. As long as there are sufficient funds, a minimum amount min_withdraw is withdrawn. More can be withdrawn if the current real growth rate supports a larger amount. The parameters below allow the user to control how much can be withdrawn as a function of real growth rate (percent_of_real_growth) and how much volitility in the withdraw amount is tolerated (volitility). The withdraw amount may also be decreased based on alternative income (e.g., social security, pension).

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

start_age = 67: the age at which withdraws begin

min_withdraw = 1000: the minimum withdraw amount

percent_of_real_growth = 1: the percent of real growth withdrawn. If equal to 1, the max of real growth or min_withdraw.

income_adjustment = 0.0: a value between 0 and 1 representing the reduction in withdraw_amount relative to other income (e.g., social security, pension, etc). 1 indicates all income is subtracted from withdraw_amount.

volitility = .1: a value greater than zero which controls the variability in withdraw amount. The standard deviation is the mean withdraw × volitility

RetirementPlanners.fixed_withdraw — Function

fixed_withdraw(
    model::AbstractModel,
    t;
    withdraw_amount = 3000.0,
    start_age = 67.0,
    income_adjustment = 0.0
)

Withdraw a fixed amount from investments per time step once retirement starts.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

withdraw_amount = 3000.0: the amount withdrawn from investments per time step

start_age = 67.0: the age at which withdraws begin

income_adjustment = 0.0: a value between 0 and 1 representing the reduction in withdraw_amount relative to other income (e.g., social security, pension, etc). 1 indicates all income is subtracted from withdraw_amount.

RetirementPlanners.variable_withdraw — Function

variable_withdraw(
        model::AbstractModel,
        t;
        start_age = 67, 
        distribution = Normal(2500, 500),
        income_adjustment = 0.0
)

Withdraw a variable amount from investments per time step once retirement starts using a specifed distribution.

Arguments

model::AbstractModel: an abstract Model object

t: current time of simulation in years

Keywords

start_age = 67: the age at which withdraws begin

distribution = Normal(2500, 500): the distribution of withdraws per time step

income_adjustment = 0.0: a value between 0 and 1 representing the reduction in withdraw_amount relative to other income (e.g., social security, pension, etc). 1 indicates all income is subtracted from withdraw_amount.